% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_backnow.R
\name{run_backnow_parallel}
\alias{run_backnow_parallel}
\title{Title}
\usage{
run_backnow_parallel(
  cl,
  input,
  MAX_ITER = 2000,
  n_chains = 3,
  norm_sigma = NULL,
  sip = NULL,
  NB_maxdelay = NULL,
  NB_size = NULL,
  printProgress = 0,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
